test(meetings): prove the panel costs the document no height
The rail's E2E asserted the ladder. The panel's asserts the claim the rail got
wrong: main's height and top are measured before and after opening, and the
panel must start where the document column ends.

Three earlier assertions on this feature passed against suggestions nobody
could see — waitForExist is satisfied by an invisible node, childElementCount
by zero-height children, 'taller than rest' by an unrelated row — so the
suggestion check measures paint: width, height, opacity, visibility and
viewport containment.

Also asserts what makes this a panel rather than a fourth tab: the draft
survives switching between notes, transcript and summary, because chat is a
lens over all three.

Captures a dense frame sequence through the whole interaction so the PR can
show the real thing rather than a mockup.
